APPDATA/requirements.txt
Flask==2.2.5
boto3==1.28.9
pymysql==1.1.0
python-dotenv==1.0.0
minio==7.2.18